Tentative update, unchecked(!!!!!), subject to change.
Created UserManager, User (temporary), modified Event to be Serializable.

TODO:
Modify EventManager implementation (needs to deserialize input on startup, and needs a serialize events method)
Write tests (!!!!! I'm not 100% sure my code works, which is why I commented so much on it to make it easier to check).
Commands

Discussion:
1. I'm not in favor of adding "past events" to our current implementation. I don't believe we've even finished implementing date/time to our existing classes yet.
2. Is it be possible for System to hold an "active user"? Else I could also make an instance variable for UserManager to hold the current active User, but it feels strange that way. Unsure.

IF WORKING ON USER IMPLEMENTATION, PLEASE PUSH TO THIS BRANCH; THIS IS A MAJOR EXTENSION AND OUR EXISTING TEST CASES OR SYSTEM MAY BREAK IF WE PUSH THIS TO MAIN BEFORE MAKING SURE IT WORKS
I would rather not rollback changes later on.